requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-haystack==2.6.0
django-toolbelt==0.0.1
gunicorn==19.7.1
jieba==0.38
mongoengine==0.11.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.1
PyJWT==1.4.2
pymongo==3.4.0
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==1.2.0
social-auth-storage-mongoengine==1.0.0
static3==0.7.0
Whoosh==2.7.4
whitenoise-3.3.0
